Approaching functional (not quite drop in) compatibility with keynav
Added in a most of the core keynav commands. The default keyanv config works
now with slight modifications for differences in keymapping, the fact that
keynav-wayland will not run daemonized (just add a keybinding to your window
manager to start it), and I still am missing support for macros and history.

Input parsing:
Input parsing is an unholy mess. I should clean that up and decide how I
actually I want it to work. I'm not sure if I really want to emulate keyanv
exactly here because I feel like it is unnecessarily non-general (ie. it
hardcodes modifiers when they can be easily gotton from the keymap). I think as
long as I choose a consistent philosophy regarding keymaps (eg. how to parse my
and match against xkb state) it will be fine, even if it is not perfectly
consistent with original keynav or like eg. sway.